+++ src/fuzzy_asteroids/util.py
@@ -115,7 +115,7 @@ class Map:
 
 class Scenario:
     def __init__(self, num_asteroids: int = 0, asteroid_states: List[Dict[str, Any]] = None,
-                 game_map: Map = None, seed: int = None):
+                 ship_state: Dict[str, Any] = None, game_map: Map = None, seed: int = None):
         """
         Specify the starting state of the environment, including map dimensions and optional features
 
@@ -124,6 +124,7 @@ class Scenario:
 
         :param num_asteroids: Optional, Number of asteroids
         :param asteroid_states: Optional, Asteroid Starting states
+        :param ship: Optional, Ship Starting state
         :param game_map: Game Map using ``Map`` object
         :param seed: Optional seeding value to pass to random.seed() which is called before asteroid creation
         """
@@ -131,6 +132,7 @@ class Scenario:
 
         # Store Map
         self.game_map = game_map if game_map else Map()
+        self.ship_state = ship_state if ship_state else {"position": self.game_map.center, "angle": 0.0}
 
         # Store seed
         self.seed = seed